ref: Merge `canonical` module into `packagePath` module
