+++ src/utils/packagePath.ts
@@ -1,6 +1,5 @@
 import * as path from 'path';
 import { RegistryPackageType } from '../targets/registry';
-import { parseCanonical } from './canonical';
 import { ConfigurationError } from './errors';
 
 /**
@@ -52,4 +51,28 @@ export function getPackageDirPath(
       `Unknown registry package type: ${registryConfType}`
     );
   }
+}
+
+/**
+ * Parses registry canonical name to a list of registry directories
+ *
+ * Example: "npm:@sentry/browser" -> ["npm", "@sentry", "browser"]
+ *
+ * @param canonicalName Registry canonical name
+ * @returns A list of directories
+ */
+function parseCanonical(canonicalName: string): string[] {
+  const [registry, packageName] = canonicalName.split(':');
+  if (!registry || !packageName) {
+    throw new ConfigurationError(
+      `Cannot parse canonical name for the package: ${canonicalName}`
+    );
+  }
+  const packageDirs = packageName.split('/');
+  if (packageDirs.some(x => !x)) {
+    throw new ConfigurationError(
+      `Cannot parse canonical name for the package: ${canonicalName}`
+    );
+  }
+  return [registry].concat(packageName.split('/'));
 }
